feature/LW-8.Display user actions. Add logic for cancel btn
